Implement LEF output
Add the make-lef command that generates the LEF view for a memory macro.
This code makes use of a generalized "layout" module that stores
information on cell locations and pins. The GDS output may now be
changed to make use of this code as well, and should become much more
simple.