Implement binary setup time analysis
Add the tsuho characterization that follows the style of tpd. It
implements a binary search for the setup time rather than a linear
sweep. An interval of minimum and maximum setup time to be tested is
determined for each input, which is then divided into 3 sampling points
at which the propagation delay is measured. This is compared to a
baseline value that is determined during the first iteration. A
violation is indicated by a 5% tpd increase. To simplify matters and to
not rely on OCEAN, the analysis of the output data is done in Python.
The simulation is run iteratively until a certain precision, i.e.
interval width, has been reached.