Implement VHDL sim model output and make command
Add the umc65/model.vhd simulation model and implement the make-model
command. The command calls into the potstill/output/model.py file which
opens the template code, evaluates all expressions in `{...}` and
returns the result.

Add the `make` command which produces all files for a memory macro of
a specific size. This shall be the frontend with which users
communicate.