Finalize GDS output
GDS output was partially implemented in main.py. However, now that the
new Layout class is in place, layout data generation can finally be
finished. Add potstill/output/gds.py and implement the phalanx script
generation. This process couldn't be any simpler: It literally boils
down to instantiating the cells that are already listed in the Layout
structure of a macro.

Adapt a few things in the potstill.layout module to the new situation.
The cells need to keep track of the GDS structure name. The pins need to
keep track of the GDS layer index for the pin label, and also calculate
a label location based on the pin geometry.

Extend the make command to also produce GDS output.